requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
markdown2==2.3.7
MarkupSafe==1.1.0
mysql-connector-python==8.0.14
pbr==5.1.1
Pillow==5.4.1
protobuf==3.6.1
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-http-client==3.1.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
stevedore==1.30.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.3.0
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.2
mysqlclient==1.3.12